Header folder created, Logo & Navbar component moved into this folder
